[rework] Split scheduler out of probeConfig
